Network related changes
Form the code to join "concept" to a brodcast based concept, where the host avertises packets with information regarding Lobby name and port in the local network via UDP, all players which want to join listen on the UDP-Port and generate a list with the Lobbies found.

First steps towards an abstraction layer which should be "agnostic" to use wheter it is a client or a server (some differences may remain (broadcast is only availabe for server/host)). The abstaction layer should also "demultiplex" incoming packets depending on the MessageID: For example the "deck-service" uses message id 5,  so it can add a callback to all unsolicited messages with id 5. Some messages may not be passed towards the demux, those are messages which "wait" for specific replies and have an special callback registered.